Change way to visualize detections.
Issue #15 was caused by errors in using the `_debug` member variable, such as
forgetting to wrap access to it in `#ifdef DEBUG`. A couple of build problems were
discovered, when the `DEBUG` flag was either set or unset, resulting from such
logical errors in the code. This issue has been resolved, and the code made a
little cleaner, by refactoring the way detections are visualized.

Instead of modifying `_debug`  whenever facial features or
head poses are detected, define member functions for drawing the
most recent detections onto a copy of the original image. Detections can be
visualized even when the flag to enable debug visualizations is not set, as in
`show_head_pose.cpp`.

Rename the flag so as not to conflate the meaning of `DEBUG`, which
conventionally refers to *Debug* vs. *Release* compilation mode.
One may want to compile the package in Release mode and still be
able to visualize detections.

NOTE: visualization of numeric IDs of facial features, as well as
reprojected points has been commented out due to personal preference.

Resolves #15.